#1334 cleanup radiobutton
