+++ lib/elementWrapper/elementWrapperFactory.js
@@ -20,8 +20,8 @@ const createElementWrapper = (elementType, query, attrValuePairs, _options, ...a
       return new ButtonWrapper(selector, options, description);
     case 'TextBox':
       return new TextBoxWrapper(selector, options, description);
-    // case 'radioButton':
-    //   return new RadioButtonWrapper(get, description, getIfExists);
+    case 'RadioButton':
+      return new RadioButtonWrapper(selector, options, description);
     case 'Checkbox':
       return new CheckBoxWrapper(selector, options, description);
     // case 'dropDown':
