+++ lib/elementWrapper/radioButtonWrapper.js
@@ -1,7 +1,61 @@
 const RadioButton = require('../elements/radioButton');
 const ElementWrapper = require('./elementWrapper');
-const { firstElement } = require('./helper');
+const { firstElement, getElementGetter } = require('./helper');
+const { $function } = require('../elementSearch');
+
+function getRadioButtonElementWithLabel(label) {
+  const radioButton = [];
+  function checkAndPushElement(elem) {
+    if (
+      elem.tagName &&
+      elem.tagName.toLowerCase() == 'input' &&
+      elem.type &&
+      elem.type.toLowerCase() === 'radio'
+    ) {
+      radioButton.push(elem);
+    }
+  }
+  const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
+    return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
+  });
+  for (let matchingLabel of matchingLabels) {
+    const labelFor = matchingLabel.getAttribute('for');
+    if (labelFor) {
+      //check label with attribute for
+      const labelForElement = document.getElementById(labelFor);
+      checkAndPushElement(labelForElement);
+    } else {
+      // check child node of label tag
+      matchingLabel.childNodes.forEach((elem) => {
+        checkAndPushElement(elem);
+      });
+    }
+  }
+  //check label is inlined
+  const inputRadioElements = [...document.querySelectorAll('input[type="radio" i]')];
+  for (let inputRadioElement of inputRadioElements) {
+    if (
+      inputRadioElement.nextSibling.nodeType === Node.TEXT_NODE &&
+      inputRadioElement.nextSibling.wholeText.includes(label)
+    ) {
+      radioButton.push(inputRadioElement);
+    }
+  }
+  return radioButton;
+}
 class RadioButtonWrapper extends ElementWrapper {
+  _get = getElementGetter(
+    this.selector,
+    async () =>
+      await $function(
+        getRadioButtonElementWithLabel,
+        this.selector.label,
+        this.options.selectHiddenElements,
+      ),
+    'input[type="radio" i]',
+    this.options.selectHiddenElements,
+  );
+
   async isSelected() {
     const elem = await firstElement.apply(this);
     return await elem.isSelected();
