+++ lib/taiko.js
@@ -2658,57 +2658,7 @@ module.exports.checkBox = (attrValuePairs, _options, ...args) => {
  */
 module.exports.radioButton = (attrValuePairs, _options, ...args) => {
   validate();
-  const { selector, options } = prepareParameters(attrValuePairs, _options, ...args);
-  function getRadioButtonElementWithLabel(label) {
-    const radioButton = [];
-    function checkAndPushElement(elem) {
-      if (
-        elem.tagName &&
-        elem.tagName.toLowerCase() == 'input' &&
-        elem.type &&
-        elem.type.toLowerCase() === 'radio'
-      ) {
-        radioButton.push(elem);
-      }
-    }
-    const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
-      return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
-    });
-    for (let matchingLabel of matchingLabels) {
-      const labelFor = matchingLabel.getAttribute('for');
-      if (labelFor) {
-        //check label with attribute for
-        const labelForElement = document.getElementById(labelFor);
-        checkAndPushElement(labelForElement);
-      } else {
-        // check child node of label tag
-        matchingLabel.childNodes.forEach((elem) => {
-          checkAndPushElement(elem);
-        });
-      }
-    }
-    //check label is inlined
-    const inputRadioElements = [...document.querySelectorAll('input[type="radio" i]')];
-    for (let inputRadioElement of inputRadioElements) {
-      if (
-        inputRadioElement.nextSibling.nodeType === Node.TEXT_NODE &&
-        inputRadioElement.nextSibling.wholeText.includes(label)
-      ) {
-        radioButton.push(inputRadioElement);
-      }
-    }
-    return radioButton;
-  }
-  const get = getElementGetter(
-    selector,
-    async () =>
-      await $function(getRadioButtonElementWithLabel, selector.label, options.selectHiddenElements),
-    'input[type="radio" i]',
-    options.selectHiddenElements,
-  );
-
-  const description = desc(selector, 'label', 'Radio button');
-  return createElementWrapper(get, description, 'radioButton');
+  return createElementWrapper('RadioButton', 'label', attrValuePairs, _options, ...args);
 };
 
 /**
